requirements.txt
pyzotero
arxiv
openai
python-dotenv
requests
pypdf
markdown
